avatar: lipsync support WIP
Now we use Kokoro's phonemizer and captioned speech output to compute a phoneme
stream with timestamps linearly interpolated from the word-level timestamps.

We also already set up a phoneme-to-animator-morph map.

Now we just need to add mouth morphs to the web API for client-side control,
and then use those in `tts_speak_lipsynced`.

Syncing the phoneme stream to the actual audio playback will be interesting
(as in "we live in interesting times").