support several avatar instances simultaneously
Really, just preparing the server to be able to serve multiple clients,
so that we can move all GPU functionality to the server side.

The avatar is the only stateful feature, all the rest are REST.

Now we could run two or more avatars at once, if the GPU has enough power. :)

(At least for now, they all run on the same device.)