move raven.common.hybridir -> raven.librarian.hybridir
Belongs there, as evidenced by it needing `raven.librarian.config`.
The `raven.common` subpackage should contain utils only (nothing that
depends on app configs).

In an app, importing stuff from another app is fine, so this is ok
even after we later integrate hybridir as an intelligent search
for Raven-visualizer.