avatar: fix missing "waver2" key in emotion templates
The live animator needs all cels to be present in the emotion template, even "waver2",
which isn't directly settable in the GUI. It is used as part of the eye-waver animation,
with the animation strength controlled by the "waver1" slider in the GUI.

Also, for testing, update the "embarrassed" emotion to use the cel blending system.
Still need to update the other templates.