librarian: store chat metadata under the new revisioned data records
Currently, there is only the RAG retrieval record (key "retrieval")
that appears when "!docs" mode is enabled, but other kinds of metadata
will be added in the near future.

The minichat datastore will auto-upgrade to the new format upon loading
(and is saved in the new format at app exit).