refactor avatar GUI driver (on client side)
This receives the live avatar video from the server, and renders it in a DPG GUI.
The video is blitted to a DPG texture, which is attached to a DPG image widget.

This gives an easier API to integrate the avatar into DPG GUI apps, as well as
provides a more compact usage example that may be useful for porting to other
GUI toolkits or to other programming languages.