librarian: initial hotkey support
Some features now have a hotkey.

Hidden debug hotkeys now available also in Librarian (Ctrl+Shift+ M, R, T, L).
Particularly interesting is Ctrl+Shift+M to see the FPS of the GUI.

Also, change the "copy message to clipboard" function so that in as-is mode,
the character name is omitted. This makes it easier to copy'n'paste an old
question into the chat field.