Name some comboboxes, update the hotkey handlers accordingly
If you tag a DPG widget when creating it, the return value of the `dpg.add_xxx`
call is the tag.

If you don't, the return value is the ID number.

We should compare the same kind of things in the hotkey handler.
It's better for debugging to name widgets than not, so the hotkey
handlers have been changed to use tags.