librarian: fade out the data eyes effect when it ends
Maybe consider later whether the fadeout code belongs in the client or in
the server. Right now it's like this, because the effect is a cel animation
directly controlled by a cel blend override, and thus any changes to the
blend value affect the blend strength immediately.

We can't easily interpolate toward a target strength in the server (like
the lipsync morphs do), because this is an *animation*, walking a cel cycle
at full strength. We simply don't have a "current strength" stored anywhere,
but we only have the target strength.

Well, at least this solution lets us configure the fadeout duration separately.