avatar_controller: add TTS idle event
Although we don't use it right now.

I was going to port `settings_editor` to use the controller, too,
but turns out that wasn't a good idea, since the recording feature
needs to capture the *full* speech audio at once, not one sentence
at a time.

Maybe this could be done by decoding the mp3s and concatenating
the results, but that loses quality. Should look into using FLAC
instead.