add tool-calling enable/disable toggle
This can be used to turn web search off temporarily.

Required some refactoring; we now dynamically inject the legacy model
tool-calling instructions in `invoke`, as well as provide or strip
the tool specifications depending on `tools_enabled`.

The feature is available both in the Raven-librarian GUI as well as
in the Raven-minichat commmand-line app (new command "!tools").