bgtask: add ManagedTask, auto-handles status box and entrypoint lock
These are auto-created per unique category tag.

This is a bit more centralized from a client viewpoint than creating those
manually and passing them in to `make_managed_task`.

The downside is that once the app no longer needs tasks of a given category,
the category's box and lock are left lying around. In practice this is not
a concern, though.